requirements.txt
gymnasium>=0.29.1
pydantic>=2.7.0
numpy>=1.26.0
pytest>=8.0.0
gradio>=4.44.0
matplotlib>=3.8.0
torch>=2.2.0
openai>=1.30.0
transformers>=4.44.0
trl>=0.11.0
peft>=0.12.0
accelerate>=0.33.0
datasets>=2.20.0
huggingface_hub>=0.24.0
python-dotenv>=1.0.1
pandas>=2.2.0
